// lib/types/enums.ts
export enum MediaType {
  IMAGE = "IMAGE",
  VIDEO = "VIDEO",
}

export enum NotificationType {
  LIKE = "LIKE",
  COMMENT = "COMMENT",
  FOLLOW = "FOLLOW",
  MESSAGE_REQUEST = "MESSAGE_REQUEST",
  MESSAGE_ACCEPTED = "MESSAGE_ACCEPTED",
  MESSAGE = "MESSAGE",
}

export enum MessageRequestStatus {
  PENDING = "PENDING",
  ACCEPTED = "ACCEPTED",
  DECLINED = "DECLINED",
}